SongTrackRendererのテストの型と関数の定義を別ファイルに切り出し
