+++ tests/unit/lib/songTrackRenderer/index.spec.ts
@@ -1,554 +1,22 @@
 import { beforeEach, describe, expect, test } from "vitest";
-import { hash } from "../../utils";
+import { SongTrackRendererTestUtility } from "./utility";
+import { RenderingEventInfo } from "./type";
 import { resetMockMode, uuid4 } from "@/helpers/random";
-import { createOpenAPIEngineMock } from "@/mock/engineMock";
-import {
-  PhraseForRender,
-  SongTrackRenderer,
-  SongTrackRenderingEvent,
-  SongTrackRenderingResult,
-} from "@/sing/songTrackRendering";
-import { EngineId, NoteId, StyleId, TrackId } from "@/type/preload";
-import {
-  createDefaultTempo,
-  createDefaultTrack,
-  DEFAULT_TPQN,
-} from "@/sing/domain";
-import { getOverlappingNoteIds } from "@/sing/storeHelper";
-import {
-  EditorFrameAudioQueryKey,
-  Note,
-  PhraseKey,
-  Singer,
-  SingingPitchKey,
-  SingingVoice,
-  SingingVoiceKey,
-  SingingVolumeKey,
-  Track,
-} from "@/store/type";
-import { calculateHash, getLast } from "@/sing/utility";
-import { ExhaustiveError, UnreachableError } from "@/type/utility";
+import { EngineId, StyleId, TrackId } from "@/type/preload";
+import { createDefaultTempo, DEFAULT_TPQN } from "@/sing/domain";
+import { UnreachableError } from "@/type/utility";
 import { getOrThrow } from "@/helpers/mapHelper";
 
-/**
- * テスト検証用のフレーズ情報。
- */
-type PhraseInfo = {
-  readonly notes: Note[];
-  readonly startTime: number;
-  readonly trackId: TrackId;
-  readonly queryKey?: EditorFrameAudioQueryKey;
-  readonly queryHash?: string;
-  readonly singingPitchKey?: SingingPitchKey;
-  readonly singingPitchHash?: string;
-  readonly singingVolumeKey?: SingingVolumeKey;
-  readonly singingVolumeHash?: string;
-  readonly singingVoiceKey?: SingingVoiceKey;
-  readonly singingVoiceHash?: string;
+const constants = {
+  tpqn: DEFAULT_TPQN,
+  tempos: [createDefaultTempo(0)],
+  engineId: EngineId("mock"),
+  singingTeacherStyleId: StyleId(0),
+  singerStyleId: StyleId(0),
+  frameRate: 93.75,
 };
 
-/**
- * フレーズの範囲情報（開始・終了tick）。
- */
-type PhraseRangeInfo = {
-  readonly startTicks: number;
-  readonly endTicks: number;
-};
-
-/**
- * テスト検証用の PhrasesGeneratedEvent 情報。
- */
-type PhrasesGeneratedEventInfo = {
-  readonly type: "phrasesGenerated";
-  readonly phraseInfos: Map<PhraseKey, PhraseInfo>;
-};
-
-/**
- * テスト検証用の CacheLoadedEvent 情報。
- */
-type CacheLoadedEventInfo = {
-  readonly type: "cacheLoaded";
-  readonly phraseInfos: Map<PhraseKey, PhraseInfo>;
-};
-
-/**
- * テスト検証用の PhraseRenderingStartedEvent 情報。
- */
-type PhraseRenderingStartedEventInfo = {
-  readonly type: "phraseRenderingStarted";
-  readonly phraseKey: PhraseKey;
-};
-
-/**
- * テスト検証用の QueryGenerationCompleteEvent 情報。
- */
-type QueryGenerationCompleteEventInfo = {
-  readonly type: "queryGenerationComplete";
-  readonly phraseKey: PhraseKey;
-  readonly queryKey: EditorFrameAudioQueryKey;
-  readonly queryHash: string;
-};
-
-/**
- * テスト検証用の PitchGenerationCompleteEvent 情報。
- */
-type PitchGenerationCompleteEventInfo = {
-  readonly type: "pitchGenerationComplete";
-  readonly phraseKey: PhraseKey;
-  readonly singingPitchKey: SingingPitchKey;
-  readonly singingPitchHash: string;
-};
-
-/**
- * テスト検証用の VolumeGenerationCompleteEvent 情報。
- */
-type VolumeGenerationCompleteEventInfo = {
-  readonly type: "volumeGenerationComplete";
-  readonly phraseKey: PhraseKey;
-  readonly singingVolumeKey: SingingVolumeKey;
-  readonly singingVolumeHash: string;
-};
-
-/**
- * テスト検証用の VoiceSynthesisCompleteEvent 情報。
- */
-type VoiceSynthesisCompleteEventInfo = {
-  readonly type: "voiceSynthesisComplete";
-  readonly phraseKey: PhraseKey;
-  readonly singingVoiceKey: SingingVoiceKey;
-  readonly singingVoiceHash: string;
-};
-
-/**
- * テスト検証用の PhraseRenderingCompleteEvent 情報。
- */
-type PhraseRenderingCompleteEventInfo = {
-  readonly type: "phraseRenderingComplete";
-  readonly phraseKey: PhraseKey;
-  readonly phraseInfo: PhraseInfo;
-};
-
-/**
- * テスト検証用の PhraseRenderingErrorEvent 情報。
- */
-type PhraseRenderingErrorEventInfo = {
-  readonly type: "phraseRenderingError";
-  readonly phraseKey: PhraseKey;
-  readonly error: unknown;
-};
-
-type RenderingEventInfo =
-  | PhrasesGeneratedEventInfo
-  | CacheLoadedEventInfo
-  | PhraseRenderingStartedEventInfo
-  | QueryGenerationCompleteEventInfo
-  | PitchGenerationCompleteEventInfo
-  | VolumeGenerationCompleteEventInfo
-  | VoiceSynthesisCompleteEventInfo
-  | PhraseRenderingCompleteEventInfo
-  | PhraseRenderingErrorEventInfo;
-
-/**
- * テスト検証用のレンダリング結果情報。
- */
-type RenderingResultInfo =
-  | {
-      readonly type: "complete";
-      readonly phraseInfos: Map<PhraseKey, PhraseInfo>;
-    }
-  | {
-      readonly type: "interrupted";
-    };
-
-// テストで使用する共通の定数
-const tpqn = DEFAULT_TPQN;
-const tempos = [createDefaultTempo(0)];
-const engineId = EngineId("mock");
-const singingTeacherStyleId = StyleId(0);
-const singerStyleId = StyleId(0);
-const frameRate = 93.75;
-
-/**
- * モックエンジンAPIを使用して SongTrackRenderer のインスタンスを生成する。
- *
- * @param playheadPositionGetter 再生ヘッド位置を取得する関数。
- * @returns SongTrackRenderer のインスタンス。
- */
-const createSongTrackRendererUsingMock = ({
-  playheadPositionGetter,
-}: {
-  playheadPositionGetter: () => number;
-}) => {
-  const mock = createOpenAPIEngineMock();
-
-  const songTrackRenderer = new SongTrackRenderer({
-    config: {
-      singingTeacherStyleId,
-      lastRestDurationSeconds: 0.5,
-      fadeOutDurationSeconds: 0.15,
-      firstRestMinDurationSeconds: 0.12,
-    },
-    engineSongApi: {
-      fetchFrameAudioQuery: async (args) => {
-        const query = await mock.singFrameAudioQuery({
-          speaker: 0,
-          score: { notes: args.notes },
-        });
-        return { ...query, frameRate };
-      },
-      fetchSingFrameF0: async (args) => {
-        return await mock.singFrameF0({
-          speaker: 0,
-          bodySingFrameF0SingFrameF0Post: {
-            score: { notes: args.notes },
-            frameAudioQuery: args.query,
-          },
-        });
-      },
-      fetchSingFrameVolume: async (args) => {
-        return await mock.singFrameVolume({
-          speaker: 0,
-          bodySingFrameVolumeSingFrameVolumePost: {
-            score: { notes: args.notes },
-            frameAudioQuery: args.query,
-          },
-        });
-      },
-      frameSynthesis: async (args) => {
-        return await mock.frameSynthesis({
-          speaker: 0,
-          frameAudioQuery: args.query,
-        });
-      },
-    },
-    playheadPositionGetter,
-  });
-
-  return songTrackRenderer;
-};
-
-/**
- * テスト用のスナップショットオブジェクトを生成する。
- *
- * @param trackEntries トラックIDとトラック情報（シンガー、ノーツ）のエントリー配列。
- * @returns スナップショットオブジェクト。
- */
-const createSnapshotObject = (
-  trackEntries: [TrackId, { singer?: Singer; notes: Note[] }][],
-) => {
-  const tracks = new Map<TrackId, Track>();
-  for (const trackEntry of trackEntries) {
-    tracks.set(trackEntry[0], {
-      ...createDefaultTrack(),
-      ...trackEntry[1],
-    });
-  }
-
-  const trackOverlappingNoteIds = new Map(
-    [...tracks.entries()].map(([trackId, track]) => [
-      trackId,
-      getOverlappingNoteIds(track.notes),
-    ]),
-  );
-
-  return {
-    tpqn,
-    tempos,
-    tracks,
-    trackOverlappingNoteIds,
-    engineFrameRates: new Map([[engineId, frameRate]]),
-    editorFrameRate: frameRate,
-  };
-};
-
-/**
- * SingingVoiceオブジェクトからハッシュ値を計算する。
- *
- * @param singingVoice 歌声オブジェクト。
- * @returns ハッシュ値のPromise。
- */
-const calculateSingingVoiceHash = async (singingVoice: SingingVoice) => {
-  return await hash(await singingVoice.arrayBuffer());
-};
-
-/**
- * PhraseForRender オブジェクトをテスト検証用の PhraseInfo オブジェクトに変換する。
- * 各レンダリング成果物のハッシュ値も計算して含める。
- *
- * @param phrase PhraseForRender オブジェクト。
- * @returns PhraseInfo オブジェクトのPromise。
- */
-const toPhraseInfo = async (phrase: PhraseForRender): Promise<PhraseInfo> => {
-  return {
-    notes: phrase.notes,
-    startTime: phrase.startTime,
-    trackId: phrase.trackId,
-    queryKey: phrase.queryKey,
-    queryHash:
-      phrase.query != undefined ? await calculateHash(phrase.query) : undefined,
-    singingPitchKey: phrase.singingPitchKey,
-    singingPitchHash:
-      phrase.singingPitch != undefined
-        ? await calculateHash(phrase.singingPitch)
-        : undefined,
-    singingVolumeKey: phrase.singingVolumeKey,
-    singingVolumeHash:
-      phrase.singingVolume != undefined
-        ? await calculateHash(phrase.singingVolume)
-        : undefined,
-    singingVoiceKey: phrase.singingVoiceKey,
-    singingVoiceHash:
-      phrase.singingVoice != undefined
-        ? await calculateSingingVoiceHash(phrase.singingVoice)
-        : undefined,
-  };
-};
-
-/**
- * PhraseForRender の Map を PhraseInfo の Map に変換する。
- *
- * @param phrases PhraseForRender の Map。
- * @returns PhraseInfo の Map のPromise。
- */
-const toPhraseInfos = async (phrases: Map<PhraseKey, PhraseForRender>) => {
-  const phraseInfos = new Map<PhraseKey, PhraseInfo>();
-  for (const [phraseKey, phrase] of phrases) {
-    phraseInfos.set(phraseKey, await toPhraseInfo(phrase));
-  }
-  return phraseInfos;
-};
-
-/**
- * SongTrackRenderingEvent をテスト検証用の RenderingEventInfo に変換する。
- * イベントの種類に応じて、関連データのハッシュ値などを計算して含める。
- *
- * @param event SongTrackRenderingEvent。
- * @returns RenderingEventInfo のPromise。
- */
-const toRenderingEventInfo = async (
-  event: SongTrackRenderingEvent,
-): Promise<RenderingEventInfo> => {
-  switch (event.type) {
-    case "phrasesGenerated":
-      return {
-        type: event.type,
-        phraseInfos: await toPhraseInfos(event.phrases),
-      };
-    case "cacheLoaded":
-      return {
-        type: event.type,
-        phraseInfos: await toPhraseInfos(event.phrases),
-      };
-    case "phraseRenderingStarted":
-      return event;
-    case "queryGenerationComplete":
-      return {
-        type: event.type,
-        phraseKey: event.phraseKey,
-        queryKey: event.queryKey,
-        queryHash: await calculateHash(event.query),
-      };
-    case "pitchGenerationComplete":
-      return {
-        type: event.type,
-        phraseKey: event.phraseKey,
-        singingPitchKey: event.singingPitchKey,
-        singingPitchHash: await calculateHash(event.singingPitch),
-      };
-    case "volumeGenerationComplete":
-      return {
-        type: event.type,
-        phraseKey: event.phraseKey,
-        singingVolumeKey: event.singingVolumeKey,
-        singingVolumeHash: await calculateHash(event.singingVolume),
-      };
-    case "voiceSynthesisComplete":
-      return {
-        type: event.type,
-        phraseKey: event.phraseKey,
-        singingVoiceKey: event.singingVoiceKey,
-        singingVoiceHash: await calculateSingingVoiceHash(event.singingVoice),
-      };
-    case "phraseRenderingComplete":
-      return {
-        type: event.type,
-        phraseKey: event.phraseKey,
-        phraseInfo: await toPhraseInfo(event.phrase),
-      };
-    case "phraseRenderingError":
-      return event;
-    default:
-      throw new ExhaustiveError(event);
-  }
-};
-
-/**
- * SongTrackRenderingResult をテスト検証用の RenderingResultInfo に変換する。
- *
- * @param result SongTrackRenderingResult。
- * @returns RenderingResultInfo のPromise。
- */
-const toRenderingResultInfo = async (
-  result: SongTrackRenderingResult,
-): Promise<RenderingResultInfo> => {
-  if (result.type === "complete") {
-    return {
-      type: result.type,
-      phraseInfos: await toPhraseInfos(result.phrases),
-    };
-  } else if (result.type === "interrupted") {
-    return result;
-  } else {
-    throw new ExhaustiveError(result);
-  }
-};
-
-/**
- * 指定されたパターン番号に基づいてテスト用のノーツを生成する。
- *
- * @param patternNumber ノートパターンの番号。
- * @returns ノーツ。
- */
-const createTestNotes = (patternNumber: number) => {
-  const quarterNoteDuration = tpqn;
-
-  const patterns = new Map([
-    [0, { lyrics: "てすと", noteNumbers: [60, 62, 64] }],
-    [1, { lyrics: "いち", noteNumbers: [60, 60] }],
-    [2, { lyrics: "に", noteNumbers: [62] }],
-    [3, { lyrics: "さん", noteNumbers: [64, 64] }],
-    [4, { lyrics: "し", noteNumbers: [65] }],
-    [5, { lyrics: "ご", noteNumbers: [67] }],
-    [6, { lyrics: "ろく", noteNumbers: [69, 69] }],
-    [7, { lyrics: "なな", noteNumbers: [71, 71] }],
-    [8, { lyrics: "はち", noteNumbers: [72, 72] }],
-  ]);
-  const pattern = getOrThrow(patterns, patternNumber);
-  if (pattern.lyrics.length !== pattern.noteNumbers.length) {
-    throw new Error("The number of noteNumbers and lyrics does not match.");
-  }
-
-  const notes: Note[] = [];
-
-  for (let i = 0; i < pattern.lyrics.length; i++) {
-    notes.push({
-      id: NoteId(uuid4()),
-      position: quarterNoteDuration * i,
-      duration: quarterNoteDuration,
-      noteNumber: pattern.noteNumbers[i],
-      lyric: pattern.lyrics[i],
-    });
-  }
-
-  return notes;
-};
-
-/**
- * 不正な歌詞（クエリの生成でエラーになる歌詞）を持つテスト用ノーツを生成する。
- *
- * @returns ノーツ。
- */
-const createTestNotesWithInvalidLyrics = (): Note[] => {
-  const quarterNoteDuration = tpqn;
-
-  return [
-    {
-      id: NoteId(uuid4()),
-      position: 0,
-      duration: quarterNoteDuration,
-      noteNumber: 60,
-      lyric: "て",
-    },
-    {
-      id: NoteId(uuid4()),
-      position: quarterNoteDuration,
-      duration: quarterNoteDuration,
-      noteNumber: 62,
-      lyric: "すと",
-    },
-  ];
-};
-
-/**
- * ノートの終了位置（Tick）を取得する。
- *
- * @param note 対象のノート。
- * @returns ノートの終了位置。
- */
-const getNoteEndPos = (note: Note) => {
-  return note.position + note.duration;
-};
-
-/**
- * フレーズのノーツの配列をトラックのノーツに変換する。
- *
- * @param phraseNotesArray フレーズのノーツの配列。
- * @returns トラックのノーツ。
- */
-const toTrackNotes = (phraseNotesArray: Note[][]) => {
-  const quarterNoteDuration = tpqn;
-  const oneMeasureDuration = tpqn * 4;
-
-  let trackNotes: Note[] = [];
-  let position = 0;
-
-  // 間に休符を挟んで連結する
-  for (const phraseNotes of phraseNotesArray) {
-    for (const note of phraseNotes) {
-      trackNotes.push({ ...note, position: position + note.position });
-    }
-
-    const lastNote = getLast(trackNotes);
-    const lastNoteEndPos = getNoteEndPos(lastNote);
-
-    position = lastNoteEndPos + quarterNoteDuration;
-  }
-
-  // ノーツ全体を1小節分後ろにシフトする
-  trackNotes = trackNotes.map((note) => ({
-    ...note,
-    position: note.position + oneMeasureDuration,
-  }));
-
-  return trackNotes;
-};
-
-/**
- * トラックのノーツをフレーズのノーツの配列に変換する。
- *
- * @param trackNotes トラックのノーツ。
- * @returns フレーズのノーツの配列。
- */
-const toPhraseNotesArray = (trackNotes: Note[]) => {
-  const phraseNotesArray: Note[][] = [];
-
-  for (let i = 0; i < trackNotes.length; i++) {
-    const note = trackNotes[i];
-
-    if (i === 0 || getNoteEndPos(trackNotes[i - 1]) < note.position) {
-      phraseNotesArray.push([note]);
-    } else {
-      getLast(phraseNotesArray).push(note);
-    }
-  }
-
-  return phraseNotesArray;
-};
-
-/**
- * ノーツから各フレーズの範囲情報（開始・終了Tick）を取得する。
- *
- * @param notes 対象のノーツ。
- * @returns フレーズ範囲情報の配列。
- */
-const getPhraseRangeInfos = (notes: Note[]): PhraseRangeInfo[] => {
-  const phraseNotesArray = toPhraseNotesArray(notes);
-
-  return phraseNotesArray.map((notes) => ({
-    startTicks: notes[0].position,
-    endTicks: getNoteEndPos(getLast(notes)),
-  }));
-};
+const utility = new SongTrackRendererTestUtility(constants);
 
 beforeEach(() => {
   resetMockMode();
@@ -557,54 +25,60 @@ beforeEach(() => {
 describe("SongTrackRenderer", () => {
   test("正しいレンダリング結果が返される", async () => {
     const trackId = TrackId(uuid4());
-    const trackNotes = toTrackNotes([
-      createTestNotes(0),
-      createTestNotes(1),
-      createTestNotes(2),
+    const trackNotes = utility.toTrackNotes([
+      utility.createTestNotes(0),
+      utility.createTestNotes(1),
+      utility.createTestNotes(2),
     ]);
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => 0,
     });
 
-    const snapshot = createSnapshotObject([
+    const snapshot = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
           notes: trackNotes,
         },
       ],
     ]);
     const result = await songTrackRenderer.render(snapshot);
 
-    const renderingResultInfo = await toRenderingResultInfo(result);
+    const renderingResultInfo = await utility.toRenderingResultInfo(result);
 
     expect({ renderingResultInfo }).toMatchSnapshot();
   });
 
   test("レンダリングイベントが正しく発行される", async () => {
     const trackId = TrackId(uuid4());
-    const trackNotes = toTrackNotes([
-      createTestNotes(0),
-      createTestNotes(1),
-      createTestNotes(2),
+    const trackNotes = utility.toTrackNotes([
+      utility.createTestNotes(0),
+      utility.createTestNotes(1),
+      utility.createTestNotes(2),
     ]);
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => 0,
     });
 
     const renderingEventInfoPromises: Promise<RenderingEventInfo>[] = [];
     songTrackRenderer.addEventListener((event) => {
-      renderingEventInfoPromises.push(toRenderingEventInfo(event));
+      renderingEventInfoPromises.push(utility.toRenderingEventInfo(event));
     });
 
-    const snapshot = createSnapshotObject([
+    const snapshot = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
           notes: trackNotes,
         },
       ],
@@ -619,42 +93,48 @@ describe("SongTrackRenderer", () => {
   test("キャッシュがロードされる", async () => {
     const trackId = TrackId(uuid4());
     const phraseNotesArray = [
-      createTestNotes(0),
-      createTestNotes(1),
-      createTestNotes(2),
+      utility.createTestNotes(0),
+      utility.createTestNotes(1),
+      utility.createTestNotes(2),
     ];
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => 0,
     });
 
     const renderingEventInfoPromises: Promise<RenderingEventInfo>[] = [];
     songTrackRenderer.addEventListener((event) => {
-      renderingEventInfoPromises.push(toRenderingEventInfo(event));
+      renderingEventInfoPromises.push(utility.toRenderingEventInfo(event));
     });
 
     // 1回目のレンダリング
-    const snapshot1 = createSnapshotObject([
+    const snapshot1 = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
-          notes: toTrackNotes(phraseNotesArray),
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
+          notes: utility.toTrackNotes(phraseNotesArray),
         },
       ],
     ]);
     const result1 = await songTrackRenderer.render(snapshot1);
 
     // 2番目のフレーズのノーツを変更
-    phraseNotesArray[1] = createTestNotes(3);
+    phraseNotesArray[1] = utility.createTestNotes(3);
 
     // 2回目のレンダリング
-    const snapshot2 = createSnapshotObject([
+    const snapshot2 = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
-          notes: toTrackNotes(phraseNotesArray),
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
+          notes: utility.toTrackNotes(phraseNotesArray),
         },
       ],
     ]);
@@ -663,7 +143,7 @@ describe("SongTrackRenderer", () => {
     const renderingEventInfos = await Promise.all(renderingEventInfoPromises);
 
     // 1回目のレンダリング結果からフレーズ情報を取得
-    const resultInfo1 = await toRenderingResultInfo(result1);
+    const resultInfo1 = await utility.toRenderingResultInfo(result1);
     expect(resultInfo1.type).toEqual("complete");
     if (resultInfo1.type !== "complete") {
       throw new UnreachableError();
@@ -698,42 +178,48 @@ describe("SongTrackRenderer", () => {
   test("変更された部分（フレーズ）のみ再レンダリングされる", async () => {
     const trackId = TrackId(uuid4());
     const phraseNotesArray = [
-      createTestNotes(0),
-      createTestNotes(1),
-      createTestNotes(2),
+      utility.createTestNotes(0),
+      utility.createTestNotes(1),
+      utility.createTestNotes(2),
     ];
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => 0,
     });
 
     // 1回目のレンダリング
-    const snapshot1 = createSnapshotObject([
+    const snapshot1 = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
-          notes: toTrackNotes(phraseNotesArray),
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
+          notes: utility.toTrackNotes(phraseNotesArray),
         },
       ],
     ]);
     await songTrackRenderer.render(snapshot1);
 
     // 2番目のフレーズのノーツを変更
-    phraseNotesArray[1] = createTestNotes(3);
+    phraseNotesArray[1] = utility.createTestNotes(3);
 
     const renderingEventInfoPromises: Promise<RenderingEventInfo>[] = [];
     songTrackRenderer.addEventListener((event) => {
-      renderingEventInfoPromises.push(toRenderingEventInfo(event));
+      renderingEventInfoPromises.push(utility.toRenderingEventInfo(event));
     });
 
     // 2回目のレンダリング
-    const snapshot2 = createSnapshotObject([
+    const snapshot2 = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
-          notes: toTrackNotes(phraseNotesArray),
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
+          notes: utility.toTrackNotes(phraseNotesArray),
         },
       ],
     ]);
@@ -756,20 +242,20 @@ describe("SongTrackRenderer", () => {
 
   test("レンダリングを中断できる", async () => {
     const trackId = TrackId(uuid4());
-    const trackNotes = toTrackNotes([
-      createTestNotes(0),
-      createTestNotes(1),
-      createTestNotes(2),
+    const trackNotes = utility.toTrackNotes([
+      utility.createTestNotes(0),
+      utility.createTestNotes(1),
+      utility.createTestNotes(2),
     ]);
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => 0,
     });
 
     const renderingEventInfoPromises: Promise<RenderingEventInfo>[] = [];
     let volumeGenerationCompleteEventCount = 0;
     songTrackRenderer.addEventListener((event) => {
-      renderingEventInfoPromises.push(toRenderingEventInfo(event));
+      renderingEventInfoPromises.push(utility.toRenderingEventInfo(event));
 
       if (event.type === "volumeGenerationComplete") {
         volumeGenerationCompleteEventCount++;
@@ -780,11 +266,14 @@ describe("SongTrackRenderer", () => {
       }
     });
 
-    const snapshot = createSnapshotObject([
+    const snapshot = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
           notes: trackNotes,
         },
       ],
@@ -810,26 +299,29 @@ describe("SongTrackRenderer", () => {
 
   test("クエリの生成でエラーが発生した場合、phraseRenderingError イベントが発行され、次のフレーズのレンダリングに進む", async () => {
     const trackId = TrackId(uuid4());
-    const trackNotes = toTrackNotes([
-      createTestNotes(0),
-      createTestNotesWithInvalidLyrics(),
-      createTestNotes(2),
+    const trackNotes = utility.toTrackNotes([
+      utility.createTestNotes(0),
+      utility.createTestNotesWithInvalidLyrics(),
+      utility.createTestNotes(2),
     ]);
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => 0,
     });
 
     const renderingEventInfoPromises: Promise<RenderingEventInfo>[] = [];
     songTrackRenderer.addEventListener((event) => {
-      renderingEventInfoPromises.push(toRenderingEventInfo(event));
+      renderingEventInfoPromises.push(utility.toRenderingEventInfo(event));
     });
 
-    const snapshot = createSnapshotObject([
+    const snapshot = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
           notes: trackNotes,
         },
       ],
@@ -867,30 +359,33 @@ describe("SongTrackRenderer", () => {
   test("トラックにシンガーが割り当てられていない場合、そのトラックのフレーズはレンダリングされない", async () => {
     const trackId1 = TrackId(uuid4());
     const singer1 = undefined;
-    const trackNotes1 = toTrackNotes([
-      createTestNotes(0),
-      createTestNotes(1),
-      createTestNotes(2),
+    const trackNotes1 = utility.toTrackNotes([
+      utility.createTestNotes(0),
+      utility.createTestNotes(1),
+      utility.createTestNotes(2),
     ]);
 
     const trackId2 = TrackId(uuid4());
-    const singer2 = { engineId, styleId: singerStyleId };
-    const trackNotes2 = toTrackNotes([
-      createTestNotes(0),
-      createTestNotes(3),
-      createTestNotes(4),
+    const singer2 = {
+      engineId: constants.engineId,
+      styleId: constants.singerStyleId,
+    };
+    const trackNotes2 = utility.toTrackNotes([
+      utility.createTestNotes(0),
+      utility.createTestNotes(3),
+      utility.createTestNotes(4),
     ]);
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => 0,
     });
 
     const renderingEventInfoPromises: Promise<RenderingEventInfo>[] = [];
     songTrackRenderer.addEventListener((event) => {
-      renderingEventInfoPromises.push(toRenderingEventInfo(event));
+      renderingEventInfoPromises.push(utility.toRenderingEventInfo(event));
     });
 
-    const snapshot = createSnapshotObject([
+    const snapshot = utility.createSnapshotObject([
       [
         trackId1,
         {
@@ -909,7 +404,7 @@ describe("SongTrackRenderer", () => {
     const result = await songTrackRenderer.render(snapshot);
 
     const renderingEventInfos = await Promise.all(renderingEventInfoPromises);
-    const renderingResultInfo = await toRenderingResultInfo(result);
+    const renderingResultInfo = await utility.toRenderingResultInfo(result);
 
     expect(renderingResultInfo.type).toEqual("complete");
     if (renderingResultInfo.type !== "complete") {
@@ -936,25 +431,25 @@ describe("SongTrackRenderer", () => {
 
   test("再生ヘッド位置に近いフレーズから優先的にレンダリングされる", async () => {
     const trackId = TrackId(uuid4());
-    const trackNotes = toTrackNotes([
-      createTestNotes(0),
-      createTestNotes(1),
-      createTestNotes(2),
-      createTestNotes(3),
-      createTestNotes(4),
-      createTestNotes(5),
-      createTestNotes(6),
+    const trackNotes = utility.toTrackNotes([
+      utility.createTestNotes(0),
+      utility.createTestNotes(1),
+      utility.createTestNotes(2),
+      utility.createTestNotes(3),
+      utility.createTestNotes(4),
+      utility.createTestNotes(5),
+      utility.createTestNotes(6),
     ]);
     const phraseIndexesPointedToByPlayhead = [3, 3, 6, 6, 6, 6, 6];
     const phraseIndexesInExpectedRenderingOrder = [3, 4, 6, 0, 1, 2, 5];
-    const phraseRangeInfos = getPhraseRangeInfos(trackNotes);
+    const phraseRangeInfos = utility.getPhraseRangeInfos(trackNotes);
 
     let currentIndex = 0;
     const movePlayheadToNextPosition = () => {
       currentIndex++;
     };
 
-    const songTrackRenderer = createSongTrackRendererUsingMock({
+    const songTrackRenderer = utility.createSongTrackRendererUsingMock({
       playheadPositionGetter: () => {
         const phraseIndex = phraseIndexesPointedToByPlayhead[currentIndex];
         const phraseRangeInfo = phraseRangeInfos[phraseIndex];
@@ -967,7 +462,7 @@ describe("SongTrackRenderer", () => {
 
     const renderingEventInfoPromises: Promise<RenderingEventInfo>[] = [];
     songTrackRenderer.addEventListener((event) => {
-      renderingEventInfoPromises.push(toRenderingEventInfo(event));
+      renderingEventInfoPromises.push(utility.toRenderingEventInfo(event));
 
       // フレーズのピッチ生成が完了したら再生ヘッド位置を更新する
       if (event.type === "pitchGenerationComplete") {
@@ -975,11 +470,14 @@ describe("SongTrackRenderer", () => {
       }
     });
 
-    const snapshot = createSnapshotObject([
+    const snapshot = utility.createSnapshotObject([
       [
         trackId,
         {
-          singer: { engineId, styleId: singerStyleId },
+          singer: {
+            engineId: constants.engineId,
+            styleId: constants.singerStyleId,
+          },
           notes: trackNotes,
         },
       ],
@@ -987,7 +485,7 @@ describe("SongTrackRenderer", () => {
     const result = await songTrackRenderer.render(snapshot);
 
     const renderingEventInfos = await Promise.all(renderingEventInfoPromises);
-    const renderingResultInfo = await toRenderingResultInfo(result);
+    const renderingResultInfo = await utility.toRenderingResultInfo(result);
 
     expect(renderingResultInfo.type).toEqual("complete");
     if (renderingResultInfo.type !== "complete") {
